refactor: rename `utils` directory to `internal` directory
