+++ src/bin.mts
@@ -2,13 +2,14 @@
 
 import chalk from "chalk";
 import ora from "ora";
-import * as utils from "./utils/index.mjs";
+
+import { ArgumentsParser, formatKeywordRank } from "./internal/index.mjs";
 import { googleGetWebsiteRank, GoogleWebsiteRank } from "./google.mjs";
 
 type RankPromise = Promise<GoogleWebsiteRank | undefined>;
 
 async function run() {
-  const parser = new utils.ArgumentsParser();
+  const parser = new ArgumentsParser();
   const args = await parser.parse();
 
   const rankByKeywords: [string, RankPromise][] = [];
@@ -27,7 +28,7 @@ async function run() {
   loading.start();
   for (const [keyword, prom] of rankByKeywords) {
     loading.text = `Getting ranks of ${chalk.blueBright(keyword)} keyword...`;
-    const str = utils.formatKeywordRank(keyword, await prom);
+    const str = formatKeywordRank(keyword, await prom);
     process.stdout.write(`\r\x1b[K${str}\n`);
   }
   loading.stop();
