+++ src/internal/format.mts
